requirements.txt
fastmcp>=2.14,<3
PyMySQL>=1.1,<2
python-dotenv>=1.0,<2
